unified-games-bot/requirements.txt
aiogram>=3.0.0
python-telegram-bot>=20.0
Flask
Flask-SQLAlchemy
requests
python-dotenv
Werkzeug
gunicorn
flask-cors
aiohttp
sqlalchemy>=2.0.0 